Move the CredsBuilders and registration back to internal/bootstrap.go.
