+++ credentials/insecure/insecure.go
@@ -22,16 +22,10 @@ package insecure
 
 import (
 	"context"
-	"encoding/json"
 	"net"
 
 	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/xds/bootstrap"
 )
-
-func init() {
-	bootstrap.RegisterCredentials(&credsBuilder{})
-}
 
 // NewCredentials returns a credentials which disables transport security.
 //
@@ -101,19 +95,4 @@ func (insecureBundle) PerRPCCredentials() credentials.PerRPCCredentials {
 // TransportCredentials returns the underlying insecure transport credential.
 func (insecureBundle) TransportCredentials() credentials.TransportCredentials {
 	return NewCredentials()
-}
-
-// credsBuilder encapsulates a insecure credential that is built using a
-// JSON config.
-type credsBuilder struct{}
-
-// BuildCredsBundle returns a default insecure credential bundle. Currently the JSON
-// config is unused.
-func (i *credsBuilder) Build(_ json.RawMessage) (credentials.Bundle, error) {
-	return NewBundle(), nil
-}
-
-// Name returns the name associated with CredsBuilder i.e. "insecure".
-func (i *credsBuilder) Name() string {
-	return "insecure"
 }
