--- credentials/insecure/insecure_test.go
@@ -1,31 +1,0 @@
-/*
- *
- * Copyright 2022 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package insecure
-
-import "testing"
-
-func TestCredsBuilders(t *testing.T) {
-	i := &credsBuilder{}
-	if _, err := i.Build(nil); err != nil {
-		t.Errorf("CredsBuilder Build error = %v, want nil", err)
-	}
-
-	if got, want := i.Name(), "insecure"; got != want {
-		t.Errorf("CredsBuilder Name = %v, want %v", got, want)
-	}
-}
